command line integration files
main contains the logic for separating and interpreting arguments given at the command line, helper is mainly for use with the help command but also is used to print results of other commands (mainly the list command).

grammar.rs contains all logic, structs, and implementations of all things related to utilizing the "derive" tooling of the command line. a possible further implementation would consist of reading rules from a file and defining them, but that is not currently implemented.